+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Fixes
+
+- ref: Set sample rates to default if out of range (#1074): When setting a value `SentryOptions.sampleRate` that is not >= 0.0 and <= 1.0 the SDK sets it to the default of 1.0.
 - feat: Add trace information from scope to event capture (#1065)
 - fix: SentryOptions.tracesSampleRate default value (#1069)
 - ref: Discard unfinished spans when capturing transaction (#1066)